+++ versions.tf
@@ -1,9 +1,9 @@
 terraform {
   required_version = ">= 1.3"
   required_providers {
-    azapi = {
-      source  = "azure/azapi"
-      version = "~> 1.3"
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = "~> 3.39"
     }
     tls = {
       source  = "hashicorp/tls"