+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Breaking
+  * AZ-1050: Use the `azurerm_storage_account_local_user` resource instead of the AzAPI resource
+
 # v7.0.0 - 2023-03-03
 
 Added